Move common helpers to utils
