+++ utils.py
@@ -2,6 +2,8 @@
 """汎用的なユーティリティ関数をまとめたモジュール"""
 
 from typing import Sequence, List
+import random
+import collections
 import numpy as np
 
 
@@ -53,3 +55,65 @@ def opponent_player(player: int) -> int:
     """与えられたプレイヤーID(1 or 2)の相手プレイヤーIDを返す"""
     # 1なら2、2なら1を返すだけのシンプルな関数
     return 2 if player == 1 else 1
+
+# ------------------------------------------------------------
+# 盤面関連のユーティリティ
+# ------------------------------------------------------------
+
+def get_valid_actions(obs: np.ndarray, env) -> list[int]:
+    """盤面から着手可能な action を列挙する"""
+    board_size = obs.shape[0]
+    empty_positions = np.argwhere(obs == 0)
+    valid_actions: list[int] = []
+    for x, y in empty_positions:
+        ix = int(x)
+        iy = int(y)
+        if env.can_place_stone(ix, iy):
+            valid_actions.append(env.coord_to_action(ix, iy))
+    return valid_actions
+
+
+def mask_probabilities(probs: np.ndarray, valid_actions: list[int]) -> np.ndarray:
+    """無効手の確率を0にして正規化した配列を返す"""
+    masked = np.zeros_like(probs)
+    for a in valid_actions:
+        masked[a] = probs[a]
+    total = masked.sum()
+    if total > 0.0:
+        masked /= total
+    return masked
+
+
+def mask_q_values(q_values: np.ndarray, valid_actions: list[int], invalid_value: float = -1e9) -> np.ndarray:
+    """無効手のQ値を ``invalid_value`` で置き換える"""
+    masked = q_values.copy()
+    mask = np.ones_like(masked, dtype=bool)
+    for a in valid_actions:
+        mask[a] = False
+    masked[mask] = invalid_value
+    return masked
+
+
+class ReplayBuffer:
+    """経験再生用のシンプルなバッファ"""
+
+    def __init__(self, capacity: int = 10000) -> None:
+        self.buffer: collections.deque = collections.deque(maxlen=capacity)
+
+    def push(self, s: np.ndarray, a: int, r: float, s_next: np.ndarray, done: bool) -> None:
+        """1ステップ分の遷移を保存"""
+        self.buffer.append((s, a, r, s_next, done))
+
+    def sample(self, batch_size: int):
+        """ランダムに ``batch_size`` 件取り出し NumPy 配列として返す"""
+        batch = random.sample(self.buffer, batch_size)
+        s, a, r, s_next, d = zip(*batch)
+        states = np.stack(s).astype(np.float32)
+        next_states = np.stack(s_next).astype(np.float32)
+        actions = np.array(a, dtype=np.int64)
+        rewards = np.array(r, dtype=np.float32)
+        dones = np.array(d, dtype=np.float32)
+        return states, actions, rewards, next_states, dones
+
+    def __len__(self) -> int:
+        return len(self.buffer)
